Game

Documentation

data Colour

This type enumerates the different colours of lambdas that arrive at the Functor Factory.

Constructors

Purple

Red

Green

Blue

Yellow

data Cell

Cells are either empty or contain a lambda of some colour.

Constructors

Empty

Lambda Colour

type Conveyor = [[Cell]]

The conveyor belt is represented as a list of lists of cells.

type Pos = (Int, Int)

Represents a position on the conveyor belt where the first component is the coordinate on the x-axis and the second component is the coordinate on the y-axis.

conveyorWidth :: Int

conveyorWidth is the width of the conveyor belt.

conveyorHeight :: Int

conveyorHeight is the height of the conveyor belt.

lambdaLimit :: Int

lambdaLimit is the maximum number of lambdas that will arrive at the factory after the initial delivery.

profit :: [Pos] -> Int

profit positions calculates the profit for positions-many lambdas.

isOutsideBounds :: Pos -> Bool

isOutsideBounds pos determines whether pos is outside the boundaries of the conveyor belt and evaluates to True if that is the case or False otherwise.

isEmpty :: Conveyor -> Bool

isEmpty conveyor determines whether all cells on the conveyor are empty or not.

makeEmpty :: Conveyor -> [Pos] -> Conveyor

makeEmpty conveyor positions should turn all cells indicated by the given positions to empty cells.

advanceConveyor :: Conveyor -> Conveyor

advanceConveyor conveyor should move all lambdas on conveyor as close to the end of the conveyor as possible.

fill :: Conveyor -> [Cell] -> Conveyor

fill conveyor lambdas should add new lambdas the conveyor.

nextMove :: Conveyor -> [Pos]

nextMove conveyor should determine the next cluster on conveyor to process with the goal of maximising the Functor Factory's profit.